create function to make war signup collector
